MAJOR: dat specification has been reworked in python
Dat specification now resides as python file instead of config file. This has the advantage that python features and codes might be used directly when parsing/using this and it reduces a large overhead (config parsing started to take upwards to a second when reloading configs, now it's much faster).

The structure of the specification is mostly the same, except that several subfields are now replaced by classes for the cases where they only have specific values; this means static attributes can now by accessed without the x['x'] syntax.

Example:
>>specification['Mods.dat'].fields['Id'].type
'ref|string'

The classes also support a as_dict() method that converts them into pure dictionary for the purpose of exporting it to json (i.e. with the json exporter or otherwise)

The specification code now resides in it's own section and was primarily moved out of mods dat. Loading of custom or default specification files can be done with PyPoE.poe.file.specification.load, setting the default specification for the dat resides in PyPoE.poe.dat.set_default_specification