WIP Wiki export in other languages (+ language config option)
Only languages that have a subfile in the current client are supported the option.
Wikis supported are russian and french
Currently only implements rather basic changes to the files used in the export. A lot of the language specific things have not be implemented yet, in particular changing of pages based on name conflicts and automatic interwiki linking.